+++ Assets/Scripts/Interfaces/IWeapon.cs
@@ -7,9 +7,9 @@ public interface IWeapon
 {
 	string Alias { get; set; }
 	WeaponType WeaponType { get; set; }
-	int FireSpeed { get; set; }
-	int ReloadSpeed { get; set; }
-	int CritAttack { get; set; }
-	int BaseAttack { get; set; }
+	float FireSpeed { get; set; }
+	float ReloadSpeed { get; set; }
+	float CritAttack { get; set; }
+	float BaseAttack { get; set; }
 	BulletType BulletType { get; set; }
 }
