+++ Assets/Scripts/Models/Weapon.cs
@@ -7,9 +7,9 @@ public class Weapon : IWeapon
 {
 	public string Alias { get; set; }
 	public WeaponType WeaponType { get; set; }
-	public int FireSpeed { get; set; }
-	public int ReloadSpeed { get; set; }
-	public int CritAttack { get; set; }
-	public int BaseAttack { get; set; }
+	public float FireSpeed { get; set; }
+	public float ReloadSpeed { get; set; }
+	public float CritAttack { get; set; }
+	public float BaseAttack { get; set; }
 	public BulletType BulletType { get; set; }
 }
